refactor(tool-classifier): hoist collect_runs/apply_runs to string_summarizer; rustfmt
- collect_runs and apply_runs were byte-for-byte identical in
  cargo.rs and nix.rs. Move both to string_summarizer.rs and
  re-import — single source of truth, no behaviour change
  (cursor[bot] PR #309 r3213267182).
- rustfmt drift in compose_recovery.rs and sandbox-server main.rs.
